requirements.txt
GitPython
Pillow
torch
torchvision
zhipuai